Feat: Allow Lua packages to be preloaded
lege_engine_preload() parses the Lua code into a function, and attaches
it to a C closure, which is then set as package.preload[name] where name
is the chunk name atm. The closure checks that the name passed to it
matches the name of the preloaded chunk, then runs the chunk, hense
loading the package.