Feat[cli]: Implement "lege run" subcommand
This reads a "project.lua", retrieves the appropriate metadata,
constructs a lege_engine_t based on that metadata, preloads any defined
modules, loads the entrypoint module, and runs the engine, without
compiling an executable.

Files that are preloaded must be referred to by there exact file name.
This is a bug, and will be fixed.