Feat: Use hedley for compiler attributes / builtins
Why roll our own #ifdefs when someone's already made a concise header
for it?

Meson will try finding the dependency on the system. If it fails, it
will use the fallback wrapped subproject, unless that is force-disabled
by the user, in which case, we'll use hedley.h in util/include. It's a
rediculously simple header, I don't think anyone will complain about me
bundling it. :P