Feat[util]: Add compatibility shims for newer Lua library features
We define luaL_pushnil(), lua_Unsigned, and LUAMOD_API if they're not
defined by the Lua implementation. This will make it easier to swap Lua
implementations, or to allow selecting a Lua implementation at compile
time. I think it's fairly unlikely we'll do this in future as it would
be a lot of work, but the features are nice to have anyway.