Feat: lege.task module with every_frame() function
This allows the user to specify tasks to be run every frame. Right now
we simply store and call the functions, but in future there will be
multiple types of task, and they will be spawned as coroutines so they
can yield for a frame, or perform I/O, without blocking the main loop.