feat: Execute pending tasks on each loop iteration
Pending tasks are resumed. If they complete, they're added to the `dead'
table where they will later be collected (this is done to avoid
modifying the `pending' table while iterating over it). The runtime does
not signal exit unless no pending and no blocked tasks exist.

This may need slight adjustments when we introduce awaiting on tasks, as
it currently wouldn't handle structured concurrency correctly if await()
was called.